+++ src/test/java/com/washingtonpost/arc/ans/v0_3/model/TestImage.java
@@ -26,7 +26,7 @@ public class TestImage extends AbstractANSTest<Image> {
         testJsonValidation("image-fixture-good", true);
         Image image = testClassSerialization("image-fixture-good");
         assertThat(image.getId(), is("unique ANS id"));
-        assertThat(image.getCreatedDate(), is(date("2015-06-25T09:50:50.52Z")));
+        assertThat(image.getCreatedDate(), is("2015-06-25T09:50:50.52Z"));
         assertThat(image.getCredits().size(), is(1));
         assertThat(image.getCredits().get(0).getName(), is("Ansel Adams"));
         assertThat(image.getCredits().get(0).getRole(), is("Photographer"));