+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 * Adding raw_html type
 * Adding blockquote and list types
 * Making Image.height and Image.widht Integers (instead of ints) so unset JSON values will not result in deserialized "0" values
+* Changing lastUpdatedDate and createdDate from java.util.Date to Strings of enforced RFC3339 format
 
 ## 0.3.2 2015/10/08